test532-gae_datastore/requirements.txt
sanic==0.7.0
aiohttp==3.3.2
aiodns==1.1.1
google-cloud-datastore==1.6.0
